requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
ape-ganache==0.7.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==2.4.1
async-timeout==5.0.1
attrs==25.3.0
base58==1.0.3
bitarray==3.4.0
cached-property==2.0.1
cchecksum==0.2.6
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
ckzg==1.0.2
click==8.1.8
colorama==0.4.6
crispy-bootstrap5==2023.10
cryptography==44.0.3
cssselect2==0.8.0
cytoolz==1.0.1
dataclassy==0.11.1
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
Django==4.2.20
django-allauth==65.7.0
django-appconf==1.1.0
django-crispy-forms==2.0
django-select2==8.0.0
djangorestframework==3.14.0
eip712==0.2.11
et_xmlfile==2.0.0
eth-account==0.11.3
eth-ape==0.8.32
eth-bloom==3.1.0
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-pydantic-types==0.1.3
eth-rlp==1.0.1
eth-tester==0.11.0b2
eth-typing==3.5.2
eth-utils==2.3.2
eth_abi==5.2.0
ethpm-types==0.6.25
eval_type_backport==0.2.2
evm-trace==0.2.4
evmchains==0.1.4
exceptiongroup==1.2.2
executing==2.2.0
frozenlist==1.6.0
greenlet==3.2.2
hexbytes==0.3.1
html5lib==1.1
idna==3.10
ijson==3.4.0
iniconfig==2.1.0
ipython==8.18.1
jedi==0.19.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lazyasd==0.1.4
lru-dict==1.2.0
lxml==5.4.0
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
morphys==1.0
msgspec==0.19.0
multidict==6.4.3
numpy==1.26.4
oscrypto==1.3.0
packaging==23.2
pandas==2.2.3
parsimonious==0.10.0
parso==0.8.4
Pillow==10.0.0
pluggy==1.5.0
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.30.2
pure_eval==0.2.3
py-cid==0.3.0
py-ecc==8.0.0
py-evm==0.10.1b1
py-geth==5.6.0
py-multibase==1.0.3
py-multicodec==0.2.1
py-multihash==0.2.3
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyGithub==1.59.1
Pygments==2.19.1
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
PyJWT==2.10.1
PyNaCl==1.5.0
PyPDF3==1.0.6
pytest==8.3.5
python-baseconv==1.2.2
python-bidi==0.6.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2025.2
pyunormalize==16.0.0
pywin32==310
PyYAML==6.0.2
qrcode==8.2
referencing==0.36.2
regex==2024.11.6
reportlab==4.0.4
requests==2.32.3
rich==13.9.4
rlp==4.1.0
rpds-py==0.24.0
safe-pysha3==1.0.4
semantic-version==2.10.0
six==1.17.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.40
sqlparse==0.5.3
stack-data==0.6.3
svglib==1.5.1
tinycss2==1.4.0
toml==0.10.2
tomli==2.2.1
toolz==1.0.0
tqdm==4.67.1
traitlets==5.14.3
trie==3.1.0
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.4.0
varint==1.0.2
watchdog==3.0.0
wcwidth==0.2.13
web3==6.20.1
webencodings==0.5.1
websockets==13.1
wrapt==1.17.2
xhtml2pdf==0.2.8
XlsxWriter==3.2.3
yarl==1.20.0
